len(s) returns ints, let's go with ints

+++ internal/reader/locking.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 )
 
-const LOCK_SECTION_SIZE = uint64(1024)
+const LOCK_SECTION_SIZE = 1024
 
 func createLocks() []sync.RWMutex {
 	numLocks := runtime.NumCPU() * 4
@@ -14,13 +14,13 @@ func createLocks() []sync.RWMutex {
 	return locks
 }
 
-func (reader *ReaderImpl) getLocksForRange(startIndexInclusive uint64, endIndexInclusive uint64) []*sync.RWMutex {
+func (reader *ReaderImpl) getLocksForRange(startIndexInclusive int, endIndexInclusive int) []*sync.RWMutex {
 	if endIndexInclusive < startIndexInclusive {
 		panic(fmt.Sprintf("Can't lock negative size range %d-%d", startIndexInclusive, endIndexInclusive))
 	}
 
-	lockIndex := (startIndexInclusive / LOCK_SECTION_SIZE) % uint64(len(reader.lineLocks))
-	lastLockIndex := (endIndexInclusive / LOCK_SECTION_SIZE) % uint64(len(reader.lineLocks))
+	lockIndex := (startIndexInclusive / LOCK_SECTION_SIZE) % len(reader.lineLocks)
+	lastLockIndex := (endIndexInclusive / LOCK_SECTION_SIZE) % len(reader.lineLocks)
 
 	var locksToUse []*sync.RWMutex
 	for {
@@ -28,31 +28,31 @@ func (reader *ReaderImpl) getLocksForRange(startIndexInclusive uint64, endIndexI
 		if lockIndex == lastLockIndex {
 			break
 		}
-		lockIndex = (lockIndex + 1) % uint64(len(reader.lineLocks))
+		lockIndex = (lockIndex + 1) % len(reader.lineLocks)
 	}
 
 	return locksToUse
 }
 
-func (reader *ReaderImpl) rLock(startIndexInclusive uint64, endIndexInclusive uint64) {
+func (reader *ReaderImpl) rLock(startIndexInclusive int, endIndexInclusive int) {
 	for _, lock := range reader.getLocksForRange(startIndexInclusive, endIndexInclusive) {
 		lock.RLock()
 	}
 }
 
-func (reader *ReaderImpl) rUnlock(startIndexInclusive uint64, endIndexInclusive uint64) {
+func (reader *ReaderImpl) rUnlock(startIndexInclusive int, endIndexInclusive int) {
 	for _, lock := range reader.getLocksForRange(startIndexInclusive, endIndexInclusive) {
 		lock.RUnlock()
 	}
 }
 
-func (reader *ReaderImpl) rwLock(startIndexInclusive uint64, endIndexInclusive uint64) {
+func (reader *ReaderImpl) rwLock(startIndexInclusive int, endIndexInclusive int) {
 	for _, lock := range reader.getLocksForRange(startIndexInclusive, endIndexInclusive) {
 		lock.Lock()
 	}
 }
 
-func (reader *ReaderImpl) rwUnlock(startIndexInclusive uint64, endIndexInclusive uint64) {
+func (reader *ReaderImpl) rwUnlock(startIndexInclusive int, endIndexInclusive int) {
 	for _, lock := range reader.getLocksForRange(startIndexInclusive, endIndexInclusive) {
 		lock.Unlock()
 	}
